Align deployment module organization
