+++ web/features/deployments/create-guide/state/scoped.ts
@@ -1,0 +1,41 @@
+import {
+  autoFilledInstanceNameAtom,
+  autoFilledReleaseNameAtom,
+  dslFileAtom,
+  dslFileReadVersionAtom,
+  envVarValuesAtom,
+  instanceDescriptionAtom,
+  instanceNameAtom,
+  isCreatingDeploymentAtom,
+  isCreatingReleaseOnlyAtom,
+  manualBindingSelectionsAtom,
+  methodAtom,
+  releaseDescriptionAtom,
+  releaseNameAtom,
+  selectedAppAtom,
+  selectedEnvironmentIdAtom,
+  sourceSearchTextAtom,
+  stepAtom,
+  submissionUnsupportedDslNodesAtom,
+} from './primitives'
+
+export const createDeploymentGuideScopedAtoms = [
+  stepAtom,
+  methodAtom,
+  sourceSearchTextAtom,
+  selectedAppAtom,
+  dslFileAtom,
+  dslFileReadVersionAtom,
+  instanceNameAtom,
+  instanceDescriptionAtom,
+  releaseNameAtom,
+  releaseDescriptionAtom,
+  autoFilledInstanceNameAtom,
+  autoFilledReleaseNameAtom,
+  selectedEnvironmentIdAtom,
+  manualBindingSelectionsAtom,
+  envVarValuesAtom,
+  submissionUnsupportedDslNodesAtom,
+  isCreatingDeploymentAtom,
+  isCreatingReleaseOnlyAtom,
+]
