+++ web/features/deployments/create-guide/state/types.ts
@@ -1,0 +1,5 @@
+import type { App, AppModeEnum } from '@/types/app'
+
+export type GuideMethod = 'bindApp' | 'importDsl'
+export type GuideStep = 'source' | 'release' | 'target'
+export type WorkflowSourceApp = App & { mode: Extract<AppModeEnum, 'workflow'> }
