+++ web/features/deployments/detail/access/README.md
@@ -4,9 +4,11 @@ Access route composition for webapp channels and environment permission sections
 
 ## Internal Modules
 
-| Module        | Why this module uses it                                            |
-| ------------- | ------------------------------------------------------------------ |
-| `route-state` | Reads the route app instance identity for access settings queries. |
+| Module                      | Why this module uses it                                            |
+| --------------------------- | ------------------------------------------------------------------ |
+| `route-state`               | Reads the route app instance identity for access settings queries. |
+| `detail/access/channels`    | Renders access channel endpoint settings.                          |
+| `detail/access/permissions` | Renders environment access policy settings.                        |
 
 ## External Modules
