+++ web/features/deployments/detail/access/permissions/README.md
@@ -4,16 +4,15 @@ Environment permission section for viewing and editing access policies on deploy
 
 ## Internal Modules
 
-| Module                          | Why this module uses it                                                       |
-| ------------------------------- | ----------------------------------------------------------------------------- |
-| `route-state`                   | Reads the route app instance identity for access-policy mutations.            |
-| `shared/components/empty-state` | Reuses deployment empty/error state presentation for environment permissions. |
-| `shared/components/section`     | Reuses the detail section frame for permissions content.                      |
-| `detail/access/state`           | Reads access route query data for environment access policies.                |
+| Module                                              | Why this module uses it                                                        |
+| --------------------------------------------------- | ------------------------------------------------------------------------------ |
+| `shared`                                            | Reuses shared deployment domain rules, UI primitives, hooks, or local helpers. |
+| `route-state`                                       | Reads the route app instance identity for access-policy mutations.             |
+| `detail/access`                                     | Reads access route query data for environment access policies.                 |
+| `detail/access/permissions/access-subject-selector` | Selects specific members and groups for access policies.                       |
 
 ## External Modules
 
-| Module                         | Why this module uses it                                                       |
-| ------------------------------ | ----------------------------------------------------------------------------- |
-| `app/components/base/skeleton` | Reuses skeleton primitives for permission loading states.                     |
-| `models/access-control`        | Reuses app access-control enums for policy conversion and permission dialogs. |
+| Module                         | Why this module uses it                                   |
+| ------------------------------ | --------------------------------------------------------- |
+| `app/components/base/skeleton` | Reuses skeleton primitives for permission loading states. |
