+++ web/features/deployments/detail/api-tokens/api-keys/README.md
@@ -4,12 +4,11 @@ API key creation, listing, and generated token presentation for the API tokens r
 
 ## Internal Modules
 
-| Module                           | Why this module uses it                                      |
-| -------------------------------- | ------------------------------------------------------------ |
-| `route-state`                    | Reads the route app instance identity for API key mutations. |
-| `shared/components/detail-table` | Reuses the detail table layout for API key rows.             |
-| `shared/components/endpoint`     | Reuses copy-pill presentation for generated token values.    |
-| `detail/api-tokens/table-styles` | Uses API-token-owned column class names for API key rows.    |
+| Module              | Why this module uses it                                                        |
+| ------------------- | ------------------------------------------------------------------------------ |
+| `shared`            | Reuses shared deployment domain rules, UI primitives, hooks, or local helpers. |
+| `route-state`       | Reads the route app instance identity for API key mutations.                   |
+| `detail/api-tokens` | Uses API-token-owned query state and table styles.                             |
 
 ## External Modules
