+++ web/features/deployments/detail/instances/row-actions/README.md
@@ -4,13 +4,11 @@ Row action menu and dialogs for deploying, redeploying, and undeploying an envir
 
 ## Internal Modules
 
-| Module                                  | Why this module uses it                                           |
-| --------------------------------------- | ----------------------------------------------------------------- |
-| `deploy-drawer/state`                   | Opens the deploy drawer from row-level deploy and redeploy flows. |
-| `route-state`                           | Reads the route app instance identity for deployment mutations.   |
-| `shared/components/detail-table-styles` | Reuses action trigger styles shared by detail tables.             |
-| `shared/domain/idempotency`             | Generates idempotency keys for undeploy and redeploy actions.     |
-| `shared/domain/runtime-status`          | Reuses deployment status rules for action availability.           |
+| Module          | Why this module uses it                                                        |
+| --------------- | ------------------------------------------------------------------------------ |
+| `deploy-drawer` | Opens the deploy drawer from row-level deploy and redeploy flows.              |
+| `shared`        | Reuses shared deployment domain rules, UI primitives, hooks, or local helpers. |
+| `route-state`   | Reads the route app instance identity for deployment mutations.                |
 
 ## External Modules
