+++ web/features/deployments/detail/overview/README.md
@@ -4,11 +4,13 @@ Overview route for release summary, environment status, and access summaries on 
 
 ## Internal Modules
 
-| Module                          | Why this module uses it                                                       |
-| ------------------------------- | ----------------------------------------------------------------------------- |
-| `route-state`                   | Reads the route app instance identity for overview query, links, and actions. |
-| `shared/components/empty-state` | Reuses deployment empty/error state presentation for the overview route.      |
-| `shared/domain/runtime-status`  | Reuses deployment status rules to compute overview counts.                    |
+| Module                               | Why this module uses it                                                        |
+| ------------------------------------ | ------------------------------------------------------------------------------ |
+| `shared`                             | Reuses shared deployment domain rules, UI primitives, hooks, or local helpers. |
+| `route-state`                        | Reads the route app instance identity for overview query, links, and actions.  |
+| `detail/overview/access-summary`     | Renders access and API token status summary sections.                          |
+| `detail/overview/environment-status` | Renders environment deployment status tiles.                                   |
+| `detail/overview/release-summary`    | Renders latest release summary and release creation entry point.               |
 
 ## External Modules
