+++ web/features/deployments/detail/overview/access-summary/access-status-section.tsx
@@ -9,7 +9,7 @@ import { SkeletonRectangle } from '@/app/components/base/skeleton'
 import Link from '@/next/link'
 import { deploymentRouteAppInstanceIdAtom } from '../../../route-state'
 import { DeploymentStatusBadge } from '../../../shared/ui/deployment-status-badge'
-import { OVERVIEW_CARD_CLASS_NAME, OVERVIEW_ICON_CLASS_NAME, OVERVIEW_INTERACTIVE_CARD_CLASS_NAME } from '../components/card-styles'
+import { OVERVIEW_CARD_CLASS_NAME, OVERVIEW_ICON_CLASS_NAME, OVERVIEW_INTERACTIVE_CARD_CLASS_NAME } from '../card-styles'
 
 type AccessStatusSectionProps = {
   accessChannels?: AccessChannels
