+++ web/features/deployments/detail/overview/card-styles.ts
