+++ web/features/deployments/detail/releases/README.md
@@ -4,9 +4,10 @@ Releases route for release history, release actions, deployment summaries, and r
 
 ## Internal Modules
 
-| Module        | Why this module uses it                                    |
-| ------------- | ---------------------------------------------------------- |
-| `route-state` | Reads the route app instance identity for release queries. |
+| Module                            | Why this module uses it                                    |
+| --------------------------------- | ---------------------------------------------------------- |
+| `route-state`                     | Reads the route app instance identity for release queries. |
+| `detail/releases/release-history` | Renders release history rows and pagination.               |
 
 ## External Modules
