+++ web/features/deployments/detail/releases/release-history/README.md
@@ -4,20 +4,14 @@ Release history section for rendering release rows, deployment summaries, and pa
 
 ## Internal Modules
 
-| Module                                                | Why this module uses it                                            |
-| ----------------------------------------------------- | ------------------------------------------------------------------ |
-| `shared/components/detail-table`                      | Reuses the detail table layout for desktop and mobile rows.        |
-| `shared/components/empty-state`                       | Reuses deployment empty/error state presentation for history rows. |
-| `shared/components/title-tooltip`                     | Shows full release labels when text is truncated.                  |
-| `shared/domain/release`                               | Formats release metadata shown in history rows.                    |
-| `shared/domain/runtime-status`                        | Reuses runtime status rules for deployed-to badges.                |
-| `shared/ui/deployment-status-style`                   | Reuses deployment status label mapping for deployed-to badges.     |
-| `detail/releases/release-actions/deploy-release-menu` | Renders release action menus inside release history rows.          |
-| `detail/releases/state`                               | Reads releases route query and pagination state.                   |
+| Module                            | Why this module uses it                                                        |
+| --------------------------------- | ------------------------------------------------------------------------------ |
+| `shared`                          | Reuses shared deployment domain rules, UI primitives, hooks, or local helpers. |
+| `detail/releases`                 | Reads releases route query and pagination state.                               |
+| `detail/releases/release-actions` | Renders release action menus inside release history rows.                      |
 
 ## External Modules
 
-| Module                           | Why this module uses it                                        |
-| -------------------------------- | -------------------------------------------------------------- |
-| `app/components/base/skeleton`   | Reuses skeleton primitives for release history loading states. |
-| `hooks/use-format-time-from-now` | Formats relative release timestamps in release history rows.   |
+| Module                           | Why this module uses it                                      |
+| -------------------------------- | ------------------------------------------------------------ |
+| `hooks/use-format-time-from-now` | Formats relative release timestamps in release history rows. |
